setup: package skeleton
because/__init__.py:
This creates the because/ directory in git and also marks it to Python as a
package in the normal sense (not an implicit namespace package).
This makes it discoverable as a Python package from which modules can be
imported at runtime, and makes it fit to provide in the setup.py packages arg.

setup.py, packages=["because"]:
This instructs setup.py/setuptools, and therefore `pip install`, to include the
`because` package directory in what is installed.